+++ animation_nodes/nodes/vector/vector_dot_product.py
@@ -1,14 +1,26 @@
 import bpy
-from ... base_types import AnimationNode
+from ... base_types import AnimationNode, VectorizedSocket
 
 class VectorDotProductNode(bpy.types.Node, AnimationNode):
     bl_idname = "an_VectorDotProductNode"
     bl_label = "Vector Dot Product"
 
+    useListA: VectorizedSocket.newProperty()
+    useListB: VectorizedSocket.newProperty()
+
     def create(self):
-        self.newInput("Vector", "A", "a")
-        self.newInput("Vector", "B", "b")
-        self.newOutput("Float", "Dot Product", "dotProduct")
+        self.newInput(VectorizedSocket("Vector", "useListA",
+            ("A", "a"), ("A", "a")))
+        self.newInput(VectorizedSocket("Vector", "useListB",
+            ("B", "b"), ("B", "b")))
+
+        self.newOutput(VectorizedSocket("Float", ["useListA", "useListB"],
+            ("Dot Product", "dotProduct"), ("Dot Product", "dotProducts")))
 
     def getExecutionCode(self, required):
-        return "dotProduct = a.dot(b)"
+        if self.useListA or self.useListB:
+            yield "_vA, _vB = VirtualVector3DList.createMultiple((a, (0,0,0)), (b, (0,0,0)))"
+            yield "amount = VirtualVector3DList.getMaxRealLength(_vA, _vB)"
+            yield "dotProducts = AN.nodes.vector.c_utils.calculateVectorDotProducts(amount, _vA, _vB)"
+        else:
+            yield "dotProduct = a.dot(b)"