+++ CHANGELOG.md
@@ -23,6 +23,7 @@
 - Vectorized *Mix* nodes.
 - *Mix Quaternions* node now uses slerp for mixing
 - Vectorized *Convert Angle* node
+- Vectorized *Vector Dot Product* node
 
 ## 2.2.0 (01 September 2020)